django-app/requirements.txt
Django
psycopg2-binary
djangorestframework
requests
pytest
pytest-django
pytest-factoryboy